+++ cmd/agy-tele/main.go
@@ -16,7 +16,7 @@ import (
 )
 
 var (
-	version = "1.0.16"
+	version = "1.0.18"
 )
 
 func main() {